+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * [#403](https://github.com/mozilla-rally/rally-core-addon/pull/403): Support opening the Rally control panel from the Rally website through the `open-rally` custom event.
 * [#417](https://github.com/mozilla-rally/rally-core-addon/pull/417): Add `rel="noopener noreferrer"` to `a` tags with target="_blank".
 * [#407](https://github.com/mozilla-rally/rally-core-addon/pull/407): Update the study card background parallax to be a fixed height.
-
+* [#419](https://github.com/mozilla-rally/rally-core-addon/pull/419): Update the "leave rally" and "leave this study" modal copy.
 # v0.9.0 (2021-02-09)
 
 [Full changelog](https://github.com/mozilla-rally/core-addon/compare/v0.8.0...v0.9.0)