+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 [Full changelog](https://github.com/mozilla-rally/core-addon/compare/v0.10.0...master)
 
+# v0.10.0 (2021-02-19)
+
+[Full changelog](https://github.com/mozilla-rally/core-addon/compare/v0.10.0...v0.10.0)
+
 # v0.9.0 (2021-02-09)
 
 [Full changelog](https://github.com/mozilla-rally/core-addon/compare/v0.9.0...v0.10.0)